chore(heading): export prop-types
